requirements.txt
pyspark==3.5.1
delta-spark==3.2.0
kafka-python==2.0.2
scikit-learn==1.3.2
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.1.4
numpy==1.24.4
